+++ Assets/Scripts/GUIScripts/WhileEvent.cs
@@ -5,15 +5,15 @@ using UnityEngine.UI;
 
 public class WhileEvent : MonoBehaviour {
 
-   public delegate void AddWhileCommand(string sensorType, string comparator, float value);
+   public delegate void AddWhileCommand(string sensorType, string comparator, int value);
    public AddWhileCommand command;
 
    public Dropdown sensorType;
    public Dropdown comparator;
 
    public void AddCommand() {
-      float value;
-      if (command != null && float.TryParse(GetComponent<InputField>().text, out value)) {
+      int value;
+      if (command != null && int.TryParse(GetComponent<InputField>().text, out value)) {
          command (sensorType.captionText.text, comparator.captionText.text, value);
       }
    }
